+++ docs/conf.py
@@ -1,167 +1,92 @@
-import importlib.metadata
+from pathlib import Path
+from urllib.parse import quote, urlencode
 
-import lsst_sphinx_bootstrap_theme
+from documenteer.conf.guide import *
 
-# Common links and substitutions =============================================
 
-rst_epilog = """
+def create_github_app_qs(domain="data-dev.lsst.cloud"):
+    """Create a URL query string with the GitHub App configuration.
 
-.. _mypy: http://www.mypy-lang.org
-.. _pre-commit: https://pre-commit.com
-.. _pytest: https://docs.pytest.org/en/latest/
-.. _tox: https://tox.readthedocs.io/en/latest/
-"""
+    See
+    https://docs.github.com/en/apps/creating-github-apps/setting-up-a-github-app/creating-a-github-app-using-url-parameters
 
-# Extensions =================================================================
+    Parameters
+    ----------
+    domain : `str`, optional
+        The domain name of the RSP environment. This is used to template some
+        of the GitHub App configuration.
 
-extensions = [
-    "sphinx.ext.autodoc",
-    # "sphinx.ext.autodoc.typehints",
-    "sphinx.ext.napoleon",
-    "sphinx.ext.doctest",
-    "sphinx.ext.intersphinx",
-    "sphinx.ext.todo",
-    "sphinx_automodapi.automodapi",
-    "sphinx_automodapi.smart_resolver",
-    "sphinx_click",
-    "documenteer.sphinxext",
-]
+    Returns
+    -------
+    qs : `str`
+        The query string with the GitHub App configuration.
+    """
+    parameters = [
+        ("name", f"Semaphore ({domain})"),
+        (
+            "description",
+            "Semaphore is the user notification service for the Rubin "
+            f"Science Platform ({domain}). An instance of the Semaphore app "
+            "is associated with each "
+            "[RSP environment]("
+            "https://phalanx.lsst.io/environments/index.html).",
+        ),
+        ("url", "https://semaphore.lsst.io"),
+        ("public", "false"),
+        ("webhook_active", "true"),
+        ("webhook_url", f"https://{domain}/semaphore/github/webhook"),
+        ("events[]", "push"),
+        ("events[]", "check_run"),
+        ("contents", "read"),
+        ("checks", "write"),
+    ]
+    return urlencode(parameters, quote_via=quote)
 
-# General configuration ======================================================
 
-source_suffix = ".rst"
+def format_org_url(org="lsst-sqre", domain="data-dev.lsst.cloud"):
+    """Format the URL creating the app for an organization.
 
-# The master toctree document.
-master_doc = "index"
+    Parameters
+    ----------
+    org : `str`, optional
+        The GitHub organization name where the app is created.
+    domain : `str`, optional
+        The domain name of the RSP environment. This is used to template some
+        of the GitHub App configuration.
 
-# General information about the project.
-project = "Semaphore"
-copyright = (
-    "2021 "
-    "Association of Universities for Research in Astronomy, Inc. (AURA)"
-)
-author = "LSST Data Management"
+    Returns
+    -------
+    url : `str`
+        The URL to create the GitHub App.
+    """
+    qs = create_github_app_qs(domain=domain)
+    url = f"https://github.com/organizations/{org}/settings/apps/new?{qs}"
+    return url
 
-version = importlib.metadata.metadata("semaphore").get("Version")
-release = version
 
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-exclude_patterns = ["_build", "README.rst"]
+def format_personal_url(domain="data-dev.lsst.cloud"):
+    """Format the URL creating the app for a user account.
 
-# The name of the Pygments (syntax highlighting) style to use.
-pygments_style = "sphinx"
+    Parameters
+    ----------
+    domain : `str`, optional
+        The domain name of the RSP environment. This is used to template some
+        of the GitHub App configuration.
 
-# The reST default role cross-links Python (used for this markup: `text`)
-default_role = "py:obj"
+    Returns
+    -------
+    url : `str`
+        The URL to create the GitHub App.
+    """
+    qs = create_github_app_qs(domain=domain)
+    url = f"https://github.com/settings/apps/new?{qs}"
+    return url
 
-# Intersphinx ================================================================
 
-intersphinx_mapping = {
-    "python": ("https://docs.python.org/3/", None),
-}
+# Generate the URLs and write them as files that are included as code blocks
+# in the Sphinx documentation.
+org_install_path = Path("user-guide/_github-app-url-org.txt")
+org_install_path.write_text(format_org_url())
 
-intersphinx_timeout = 10.0  # seconds
-intersphinx_cache_limit = 5  # days
-
-# Linkcheck builder ==========================================================
-
-linkcheck_retries = 2
-
-# linkcheck_ignore = [r'^https://jira.lsstcorp.org/browse/']
-
-linkcheck_timeout = 15
-
-# HTML builder ===============================================================
-
-html_baseurl = "https://semaphore.lsst.io"
-
-templates_path = [
-    "_templates",
-    lsst_sphinx_bootstrap_theme.get_html_templates_path(),
-]
-
-html_theme = "lsst_sphinx_bootstrap_theme"
-html_theme_path = [lsst_sphinx_bootstrap_theme.get_html_theme_path()]
-
-html_context = {}
-
-# Theme options are theme-specific and customize the look and feel of a theme
-# further.  For a list of options available for each theme, see the
-# documentation.
-html_theme_options = {"logotext": project}
-
-# The name for this set of Sphinx documents.  If None, it defaults to
-# "<project> v<release> documentation".
-html_title = f"{project} v{version}"
-
-# A shorter title for the navigation bar.  Default is the same as html_title.
-html_short_title = project
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = []
-
-# If true, links to the reST sources are added to the pages.
-html_show_sourcelink = False
-
-# Do not copy reST source for each page into the build
-html_copy_source = False
-
-# If false, no module index is generated.
-html_domain_indices = True
-
-# If false, no index is generated.
-html_use_index = True
-
-# API Reference ==============================================================
-
-napoleon_google_docstring = False
-napoleon_numpy_docstring = True
-napoleon_include_init_with_doc = False
-napoleon_include_private_with_doc = False
-napoleon_include_special_with_doc = True
-napoleon_use_admonition_for_examples = False
-napoleon_use_admonition_for_notes = False
-napoleon_use_admonition_for_references = False
-napoleon_use_ivar = False
-napoleon_use_keyword = True  # TODO
-napoleon_use_param = True
-napoleon_use_rtype = True
-
-autosummary_generate = True
-
-automodapi_inheritance_diagram = False
-automodapi_toctreedirnm = "api"
-automodsumm_inherited_members = False
-
-# Docstrings for classes and methods are inherited from parents.
-autodoc_inherit_docstrings = True
-
-# Class documentation should only contain the class docstring and
-# ignore the __init__ docstring, account to LSST coding standards.
-autoclass_content = "class"
-
-# Default flags for automodapi directives. Special members are dunder
-# methods.
-autodoc_default_options = {
-    "show-inheritance": False,
-    "special-members": True,
-}
-
-# Render inheritance diagrams in SVG
-graphviz_output_format = "svg"
-
-graphviz_dot_args = [
-    "-Nfontsize=10",
-    "-Nfontname=Helvetica Neue, Helvetica, Arial, sans-serif",
-    "-Efontsize=10",
-    "-Efontname=Helvetica Neue, Helvetica, Arial, sans-serif",
-    "-Gfontsize=10",
-    "-Gfontname=Helvetica Neue, Helvetica, Arial, sans-serif",
-]
-
-# TODO extension =============================================================
-
-todo_include_todos = False
+personal_install_path = Path("user-guide/_github-app-url-personal.txt")
+personal_install_path.write_text(format_personal_url())